Fix macro usage. Propagate changes.
- Fit macros to use ALBANY_ABORT(...) instead of
ALBANY_ASSERT(false, ...).

- Propagate changes from permafrost model to ice
model for fixing the failing NBC test. There is
still a subtle and weird Sacado error that only
shows in release builds, when multiplying
expressions of type double * FAD * FAD. It seems
that there is some issue with the lazy evaluation
that leads to an FPE, but it only shows if
the above expression type is assigned to another
variable, as it can be evaluated on its own
without problem. Discuss with Irina to see if
she has ideas about what the problem might be.