Add boundary and populate boundary fields
Add boundary indicator fields for faces, edges and nodes,
in addition to the existing cell one. Populate them in the
topology class, which is relatelively easy to do for nodes.
Edges are not needed for ACE and Albany lacks data layouts
for edges anyway, so we leave that as a placeholder only.

Faces are needed for Neumann BCs, but STK only deals with
elements and nodes. We have wrestled with STK to make it
generate faces for ACE application, but since NBCs are
specified by sidesets, it's unclear how to map those
to faces.

In any case, the NBCs are for the bottom of the domain,
where faces do not advance, but simply removed from the
mesh, so we'll concentrate on working on moving DBCs
for the time being.