Add moving SDBC
Needs extensive testing.

It works by adding into an SDBC nodeset all potential nodes
that could be in the boundary at some time.

It is necessary to request the creation of a
node_boundary_indicator field in the input file. Also, this
field would only be filled with valid values if Erosion
Adaptation is requested.

If these two conditions are met, the node_boundary_indicator
field is computed by the topology graph data structure.
Each node in the mesh will have a value associated with it
that indicates whether it is on a boundary.

If so, the SDBC is applied to it, otherwise it is ignored,
making it appear as if the SDBC is moving with the
boundary.