Add ability to print solution to MM and std::cout
Both the residual and Jacobian have the ability to be
printed to matrix market format or to std::cout
everytime they are evaluated, or every specified
number of evaluations.

But the solution only had the ability to be printed
at the end of the run. Fix this and make the solution
debug output behave the same way as the residual and
Jacobian debug output.