Change exodus naming for adaptation.
Due to a Paraview bug, files with erosion were not
displayed in sequence, and if a file had only one
time snapshot, the time was not displayed correctly.

A workaround is to rename the exodus ouput files
as a sequence so that they are displayed correctly,
as follows:

cube.e   -> cube.e-s.0
cube_1.e -> cube.e-s.1
cube_2.e -> cube.e-s.2

In this example, the cube.e-s.0 file is the original,
intact mesh. The subsequent files are indexed according
to how many adaptation steps have occurred in that time
period. The sequence of cube.e-s.* files will be
displayed in Paraview as a single group, showing the
progression of adaptation, and displaying the time
correctly.