Changes due to Seacas and Sacado compile errors
- Seacas: Modify CMake options to let the Trilinos
configuration process decide whether parallel netCDF
is required and present. Otherwise compilation fails.

- Sacado: There was a redesign of the FAD types that resulted
in a compilation error in the FieldFrobeniusNorm class due
to a SFINAE problem. Digging uncovered that this class and
associated helper functions are not used by anyone, so
all that code was simply removed.

There are now many tests that fail with a Kokkos related error,
but pushing this anyway since otherwise there will be no
nightly testing at all.