ACE: writing tests for sequential thermo-mechanical solver that compares thermal solution to the solution computed in the CuboidThermoMechanical3D_ThermalOnly test using Exodiff.
Note that currently the sequential thermo-mechanical solver solves the mechanics and thermal
problems separately without any coupling.  The coupling needs to be added as a next step.

Note also that the mechanics part does nothing effectively - it computes a solution of all 0s.
That needs to be switched over to the J2 Erosion material model, tested, and modified to
solve a nontrivial problem.