ACE: reworking coupler to write out new Exodus file at the end of each controller time-step, towards enabling restarts.
It looks like the output with this approach is going to be REALLY ugly,
namely with many many output files dumped.  These files can/should probably
be named using the same naming convention as the erosion, so that they can all be loaded
into paraview together, like with the erosion files.

One complication will be the case when people do not request I/O
at every time step.  It will need to be done at every controller time-step
anyway for the restarts.  Need to think about how to handle this...

Also, for some reason, the Exodus files written out during each
controller time-step contain three snapshots, not two, with the third
one corresponding to the second time + 1.  Need to figure out what is
going on there.