Fix problem with missing ACE Ice Saturation
Turns out that the resize() function left the
old contents in place, so the arrays needed to be
cleared before copying.

Still, it's odd that durung the first time step
the ice saturation is in the state manager, but
not in the second.

Since we don't really care about it other than
the ability to read it, that is OK.