Creating Explicit version of Thermal2D problem, which has a time-dept BC.
I am able to reproduce Jenn's BC "sticking" in the Exodus output file
on this problem, whose resolution can be easily controlled by changing the input file.

I will debug the issue on this problem.

This problem does not show up with SDBCs + Implicit, so it seems to have
to do with explicit steppers only.  It also shows up with different explicit steppers
(e.g. GERK), not just Forward Euler.