ACE: fixing issue in master with Jacobian registration - ACE thermo-mechanical problem runs now :).
It looks like the ice_saturation field in the J2Erosion model and MechanicsResidual is 0
which is not right.  I am sure initially in this implementation, it wasn't 0 at least in MechanicsResidual evaluator.
Need to discuss with Alejandro -
the issue could be as simple as having lower vs. upper case, or missing the '_' in the
name 'ACE Ice Saturation'.