ACE: created a branch off of this commit where reading of ice saturation was working into mechanics residual: ad2e4197522089e14931352df603a02547ff75d9 .
It still works, and I can read it in the J2Erosion model just fine too - adding print statement with
this commit to show that. So it's possible to read things in correctly (non-zeros) w/ LoadStateField.

There are other things not working in this branch with the coupling.  I think I will maybe try to
cherry-pick things from master into this branch to get everything working here, rather than try to figure
out what is different w/ the reading in this branch and master.  It seems to me this approach may
be easiest/fastest.