ACE: removing unused variable.
With this commit, I've finished cherry-picking relevant commits
from master into this branch, which fixes the ace ice saturation
issue.  Probably this branch should be merged in at this point to
replace the current master.

We can save the old master as a tag/different branch if this is of interest.