ACE: creating example that does thermal explicit coupled with mechanics implicit.
I noticed that when running explicitly, the initial thermal solution in the Exodus file
appears to be incorrect at the initial time-step in the Exodus file.  This happens with
a stand-alone thermal problem too, so it does not have to do with the coupling.
Specifically, it looks like the first time-step written to the Exodus file is off by one time-step.

I will look into/fix this later in the week.  I think one can still start trying to run things
as is to see if it runs/identify any issues e.g., with the erosion.