ACE: Adding
Initial Condition Consistency: Consistent

option to thermal input files to ensure BCs get written to the initial solution
in the Exodus output file, as this is arguably the more correct thing to
write to the output file initially.  Unfortunately it's not straightforward
to have the code set this option internally.  Also, unfortunately, the option
does not work with DBCs, only SDBCs.  Hence, test cases have been switched to use
SDBCs, which are in general preferred anyway.