Fix reference issue to NOX solver.
For some reason, if the reference to the NOX solver
is constant, it's not possible to obtain the status
of the solver (!!??).

Use const_cast to cast away the const, then it works.
This is needed to check whether NOX was successful
in solving the nonlinear system, and if not,
reduce the time step.