ACE: Rework wave pressure BC (WIP)
Cubit removes side sets that aren't a the boundary of the mesh.
This is problematic for the moving pressure BC, as there are
no side sets left when erosion happens. Thus, new side sets
need to be computed on the fly. One way to do this is to
pull the topology structure that determines the boundary
indicator fields, since it always contains up-to-date
boundary information. The plan is to use this info to
fill up the arrays that the pressure BC needs. Introduce
methods that will allow the BC to access and manipulate
the topology structure.

Other minor changes.