Reworked SurfaceElementLocking tests
After an upgrade to Fedora 34 these tests started failing.
The problem was that the SEACAS algebra program was segfaulting.
I tried to fix this by modifying the alg.i instructions, but the
segfault persisted even when I was sure there were no references
to inexistent variables, all indices correct, etc.

Upon closer inspection it turned out that the use of the algebra
program was unnecessary. Perhaps it was needed at some point in
the past to manipulate the fields of the surface elements. But
as far as I can tell, the output of the tests can be compared
directly with exodiff to the gold files. So I eliminated the
use of the algebra program altogether.