Rework adative time stepping algoritm
Rework the adative time stepping algorithm for sequential
thermomechanical coupling. In this version the logic is
very similar to the corresponding algorithm for Schwarz.
It appears to work on my limited testing, one of which
is the mini erosion test. But some corner cases may
still lurk under the surface ...

Also, this algorithm uncovered an issue in null space
utils, in that total erosion of the mesh triggers
an FPE. Introduce an ALBANY_PANIC statement to
prevent this. May need to introduce more robust
handling on meshes that have been eroded completely.