ACE: putting in infrastructure for new wave pressure NBC.
Initially I am planning to have both the old and new wave pressure NBC
in the code (in this fork) while we are verifying the new
implementation.  Some of the old wave pressure NBC input files need
modification to work with both the new and old formulations.

Still need to implement the new expression in PHAL::Neumann
but this will be easy to do.