Checking in input files for generating results in Schwarz paper.
I don't have a set of input files for each case reported in the paper
but the input files could be modified easily to reproduce all the cases.
Hopefully this is sufficient.

Missing .jou file for Ct-Hex coupling for torsion problem.

Need to try running cases to verify that they all work.