Add guard for file renaming.
The renaming of files in parallel was failing under 2
conditions: when the old and new file names were the same
and the processor count was higher than 8. The cause
for the failure of the standard rename C++ function
under these conditions is unknown (to me at least).

A simple workaround is to add a guard to the
renameParallel function when the old and new file names
are the same.