Trial fix for adaptive time step for thermal
It seems that Tempus cannot reuse the time integrator when
failure is detected in the thermo-mechanical solver
and the time step is reduced. Change the logic to destroy
the previous integrator and create a new one when this happens.
This seems to resolve this issue, but needs more testing.

There still remains the issue that this is not suffcient
for the thermal solver. Both the solver and the time
step reduction are invoked, but at some point the thermal
solver still fails with a tiny time step.