Boost <= 1.8 incompatible with Clang >=18
According to discussions online, Boost <= 1.8 has a bug that
is only triggered by Clang >= 18. RHEL9 updated to Clang 18
which triggered this bug. Downgrade to Clang 15 to work
around this for the time being, as we wait for an update
for Boost or install our own.

This also required loosening the tolerances for one of
the Schwarz tests.