Fixing failing ACE and ACI tests that happened due to changes in freezing curve definition.
The tests that were not converging had an initial temperature < 220
will lead to issues in new formulation.  The fix was raising the initial
temperature to 250.  The tests using exodiff required adding
ace_freezing_curve to gold files.  A couple of the tests had a slight
change in their mean value response, and the target value was adjusted.