Added a warning instead of an abort when reaching temperature values that cause the freezing curve gradient to overflow, because the solver often reaches these values in one of the iterations on the nonlinear system (for implicit thermal), but those solutions don't actually become the final accepted solution, and it was causing issues by aborting a process that would ultimately converge to an acceptable value
Also added a new parameter 'ACE Material Fit Truncation Ice Saturation'. If the ice saturation of an integration point is below this value, the code will adopt the Residual Elastic Modulus for the Elastic Modulus and the Soil Yield Strength for the Yield Strength. This is optional, and if not specified the code should behave as it did before.